Add files via upload
This Project focus on building screenshot tool that's compatible with any target. capable of getting the current state of the target in images format and saves it either remotely or onsite. even though there are many screenshot tools out there. I still want to know how things are created and maybe I can make some impact in the technology world